Create, edit and delete all working for medications. Validations on the form and dates in the correct format

+++ db/migrate/20191231124716_medications.rb
@@ -4,7 +4,7 @@ class Medications < ActiveRecord::Migration[6.0]
       t.string :name
       t.date :start_date
       t.date :end_date
-      t.integer :dose
+      t.float :dose
       t.text :notes
 
       t.timestamps
